--- src/v2/components/ModalHandler/Modals/LoginModal/index.js
@@ -1,15 +1,0 @@
-import { useDispatch } from "react-redux";
-import { LoginForm } from "v2";
-import { LOGIN_MODAL_CLOSE } from "v2/utils";
-import Modal from "../../Modal";
-
-function LoginModal() {
-  const dispatch = useDispatch();
-  return (
-    <Modal onClose={() => dispatch({ type: LOGIN_MODAL_CLOSE })}>
-      <LoginForm />
-    </Modal>
-  );
-}
-
-export default LoginModal;
