--- src/v2/components/ModalHandler/Modals/SignupModal/index.js
@@ -1,15 +1,0 @@
-import { useDispatch } from "react-redux";
-import { SignupForm } from "v2";
-import { SIGNUP_MODAL_CLOSE } from "v2/utils";
-import Modal from "../../Modal";
-
-function SignupModal() {
-  const dispatch = useDispatch();
-  return (
-    <Modal onClose={() => dispatch({ type: SIGNUP_MODAL_CLOSE })}>
-      <SignupForm />
-    </Modal>
-  );
-}
-
-export default SignupModal;
