+++ src/v2/components/Modals/LoginModal/index.js
@@ -1,0 +1,18 @@
+import { useDispatch, useSelector } from "react-redux";
+import { LoginForm, Modal } from "v2";
+import { LOGIN_MODAL_CLOSE } from "v2/utils";
+
+function LoginModal() {
+  const dispatch = useDispatch();
+  const { modalLogin } = useSelector((state) => state.modals);
+  return (
+    <Modal
+      isOpen={modalLogin.isVisible}
+      onClose={() => dispatch({ type: LOGIN_MODAL_CLOSE })}
+    >
+      <LoginForm />
+    </Modal>
+  );
+}
+
+export default LoginModal;
