+++ src/v2/components/Modals/index.js
@@ -1,2 +1,3 @@
+export { default as Modal } from "./Modal";
 export { default as LoginModal } from "./LoginModal";
 export { default as SignupModal } from "./SignupModal";
