+++ src/v2/components/index.js
@@ -1,6 +1,6 @@
 export * from "./Forms";
 export * from "./AuthComponents";
-export * from "./ModalHandler/Modals";
+export * from "./Modals";
 
 export { default as Navigation } from "./Navigation";
 export { default as ModalHandler } from "./ModalHandler";
